Reorganizing files a bit:   - adding an extension for constraints initialization   - adding Marks to separate concerns in NeatTipView
